+++ lib/phlex/compiler/generators/element.rb
@@ -49,11 +49,11 @@ module Phlex
 				private
 
 				def tag
-					HTML::STANDARD_ELEMENTS[@node.name] || HTML::VOID_ELEMENTS[@node.name]
+					View::STANDARD_ELEMENTS[@node.name] || View::VOID_ELEMENTS[@node.name]
 				end
 
 				def void?
-					HTML::VOID_ELEMENTS.key?(@node.name)
+					View::VOID_ELEMENTS.key?(@node.name)
 				end
 			end
 		end
