+++ test/phlex/view/naughty_business.rb
@@ -41,7 +41,7 @@ describe Phlex::View do
 		end
 	end
 
-	Phlex::HTML::EVENT_ATTRIBUTES.each_key do |event_attribute|
+	Phlex::View::EVENT_ATTRIBUTES.each_key do |event_attribute|
 		with "with naughty #{event_attribute} attribute" do
 			naughty_attributes = { event_attribute => "alert(1);" }
